✨ Refactor Home Widget service and enhance macOS compatibility
This commit decouples the `HomeWidget` logic by moving its implementation into a dedicated service class, refactors the `WeatherBloc` to use named parameters, and updates macOS project configurations for better compatibility and shared container support.

### Core Architecture & Logic
- **`HomeWidgetService`**: Extracted implementation from the interface file into `lib/services/home_widget_service_impl.dart`.
- **`WeatherBloc`**:
    - Refactored constructor to use named parameters for improved readability and maintenance.
    - Updated `WeatherInitial` state to require a `dailyForecast` object, ensuring consistent state initialization.
    - Switched from direct `HomeWidget` calls to using the `_homeWidgetService` abstraction.
- **`LocalDataSource`**:
    - Updated `getAppDirectory` to support macOS shared containers via a method channel.
    - Improved image saving logic to handle macOS similarly to iOS for widget integration.
    - Switched to `path_provider` with an alias and refined directory retrieval logic.

### macOS Configuration
- **Deployment Target**: Bumped `MACOSX_DEPLOYMENT_TARGET` from `10.15` to `11.0` in `project.pbxproj`, `Podfile`, and `Podfile.lock`.
- **Entitlements**: Added `com.apple.security.application-groups` with the `group.dmytrowidget` identifier to both Debug and Release entitlement files to enable shared data access for widgets.

### Infrastructure & Cleanup
- **Dependencies**: Updated `build_runner` to `2.12.1` and updated `pubspec.lock` accordingly.
- **Constants**: Renamed `appleAppGroupId` to `kAppleAppGroupId` for better adherence to naming conventions.
- **Tests**: Updated `weather_bloc_test.dart` and `weather_page_test.dart` to reflect `WeatherBloc` constructor changes and new state requirements.
- **UI**: Minor refactoring of conditional logic in `AboutPage` for better readability.
