+++ macos/Runner/DebugProfile.entitlements
@@ -9,8 +9,12 @@
 	<key>com.apple.security.network.server</key>
 	<true/>
 	<key>com.apple.security.network.client</key>
-    <true/>
-    <key>com.apple.security.personal-information.location</key>
-    <true/>
+	<true/>
+	<key>com.apple.security.personal-information.location</key>
+	<true/>
+	<key>com.apple.security.application-groups</key>
+	<array>
+		<string>group.dmytrowidget</string>
+	</array>
 </dict>
 </plist>
