+++ pubspec.yaml
@@ -69,7 +69,7 @@ dependencies:
 
 dev_dependencies:
   bloc_test: ^10.0.0
-  build_runner: ^2.11.1
+  build_runner: ^2.12.1
   envied_generator: ^1.3.3
   flutter_lints: ^6.0.0
   flutter_test:
